Move `Keys` to `keys.ts`
